fixing the book button

+++ lib/Pages/CartPage.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
+import 'package:firebase_auth/firebase_auth.dart'; // Import FirebaseAuth
 
 import '../component/customAppBar.dart';
 import '../component/BookingCard.dart';
@@ -34,8 +35,13 @@ class _CartPageState extends State<CartPage> {
     });
 
     try {
+      final String? uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid == null) {
+        throw Exception("User not logged in");
+      }
+
       final response = await http.get(
-        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/cart.json'),
+        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/users/$uid/cart.json'),
       );
       print("Fetched bookings response: ${response.body}");
 
@@ -71,6 +77,11 @@ class _CartPageState extends State<CartPage> {
 
   Future<void> addToCart(Map<String, dynamic> yogaClass) async {
     try {
+      final String? uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid == null) {
+        throw Exception("User not logged in");
+      }
+
       // Check if the class already exists in the cart
       final existingBooking = bookings.firstWhere(
         (booking) => booking['id'] == yogaClass['id'],
@@ -85,7 +96,7 @@ class _CartPageState extends State<CartPage> {
       }
 
       final response = await http.post(
-        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/cart.json'),
+        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/users/$uid/cart.json'),
         headers: {'Content-Type': 'application/json'},
         body: json.encode(yogaClass),
       );
@@ -110,8 +121,13 @@ class _CartPageState extends State<CartPage> {
 
   Future<void> moveToBooked(String firebaseKey, Map<String, dynamic> booking) async {
     try {
+      final String? uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid == null) {
+        throw Exception("User not logged in");
+      }
+
       final response = await http.post(
-        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/Booked.json'),
+        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/users/$uid/booked.json'),
         headers: {'Content-Type': 'application/json'},
         body: json.encode(booking),
       );
@@ -135,14 +151,15 @@ class _CartPageState extends State<CartPage> {
   }
 
   Future<void> deleteFromCart(String firebaseKey, {bool showMessage = true}) async {
-    print("Deleting booking with key: $firebaseKey");
-
     try {
+      final String? uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid == null) {
+        throw Exception("User not logged in");
+      }
+
       final response = await http.delete(
-        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/cart/$firebaseKey.json'),
+        Uri.parse('https://universal-yoga-8f236-default-rtdb.firebaseio.com/users/$uid/cart/$firebaseKey.json'),
       );
-      print("Response status: ${response.statusCode}");
-      print("Response body: ${response.body}");
 
       if (response.statusCode == 200) {
         fetchBookings();
